requirements.txt
# ===== Backend (Python) =====
flask==2.3.2
flask-cors==3.0.10
web3==6.4.0
python-dotenv==1.0.0
pymongo==4.5.0
gunicorn==20.1.0
pytest==7.3.1

# === Extra dependencies used in code ===
pillow==10.4.0
numpy==2.1.1
joblib==1.4.2
pyjwt==2.8.0
scikit-learn==1.5.2
sentence-transformers==2.2.2
transformers>=4.0.0
requests==2.31.0
beautifulsoup4==4.12.0

